requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
boto3==1.38.6
botocore==1.38.6
celery==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
dj-database-url==2.3.0
dj-rest-auth==7.0.1
Django==5.2
django-allauth==65.7.0
django-anymail==13.0
django-celery-beat==2.8.0
django-cors-headers==4.7.0
django-filter==24.3
django-push-notifications==3.2.1
django-storages==1.14.6
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.15.2
djangorestframework-gis==1.1
djangorestframework_simplejwt==5.5.0
frozenlist==1.6.0
GDAL==3.10.3
haversine==2.9.0
idna==3.10
jmespath==1.0.1
kombu==5.5.3
Markdown==3.8
multidict==6.4.3
pillow==11.2.1
prompt_toolkit==3.0.51
propcache==0.3.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
redis==5.2.1
requests==2.32.3
s3transfer==0.12.0
six==1.17.0
sqlparse==0.5.3
twilio==9.5.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
yarl==1.20.0
